requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
drf-spectacular
djangorestframework-authtoken
